Add ?? method that returns None if a block throws an NPE.
This can be used to chain references that might be null:

val result: Option[A] = ??(a.b.c)

If a or b is null, the NPE is caught and None is
returned. If c evaluates to null, None is returned.
Otherwise the result of c is wrapped in a Some.

I'm not certain this is really a good idea.